+++ CHANGELOG.md
@@ -6,3 +6,5 @@
 - [Feature: Router support request label.](https://github.com/Tencent/spring-cloud-tencent/pull/165)
 - [Feature: Support router expression label](https://github.com/Tencent/spring-cloud-tencent/pull/190)
 - [Add metadata transfer example.](https://github.com/Tencent/spring-cloud-tencent/pull/184)
+- [Feature: Support metadata router.](https://github.com/Tencent/spring-cloud-tencent/pull/191)
+- [Feature: Misc optimize metadata router.](https://github.com/Tencent/spring-cloud-tencent/pull/192)